teams-python-agno/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# AgentOps
agentops==0.2.5

# LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector DB
chromadb==0.4.18

# MongoDB
motor==3.3.2
pymongo==4.6.0

# OpenAI
openai==1.3.7

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
